+++ src/components/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CodeSandbox/code-sandbox';
